Introduce WorkflowDetails and WorkflowDetailsCache (unfinished)
